Remove dependency from validator to node (needed some clean up of quickstart)
(cherry picked from commit 15b0f1ab818483e822060793b8a1810f7b61a6d1)
