+++ src/Nethermind/Nethermind.HonestValidator.Test/TestSystem.cs
@@ -20,7 +20,6 @@ using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
 using Nethermind.BeaconNode.OApiClient;
-using Nethermind.BeaconNode.Services;
 using Nethermind.Core2.Configuration;
 using Nethermind.Core2.Crypto;
 using Nethermind.Core2.Cryptography;
