+++ src/Nethermind/Nethermind.HonestValidator.Test/ValidatorClientTest.cs
@@ -25,9 +25,7 @@ using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using Nethermind.BeaconNode;
-using Nethermind.BeaconNode.MockedStart;
 using Nethermind.BeaconNode.OApiClient;
-using Nethermind.BeaconNode.Services;
 using Nethermind.Core2;
 using Nethermind.Core2.Crypto;
 using Nethermind.Core2.Json;
