+++ src/Nethermind/Nethermind.HonestValidator/HonestValidatorWorker.cs
@@ -20,8 +20,6 @@ using System.Threading.Tasks;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
-using Nethermind.BeaconNode;
-using Nethermind.BeaconNode.Services;
 using Nethermind.Core2;
 using Nethermind.Core2.Configuration;
 using Nethermind.HonestValidator.Services;
