+++ src/Nethermind/Nethermind.HonestValidator/MockedStart/QuickStartKeyProvider.cs
@@ -21,9 +21,8 @@ using System.Numerics;
 using System.Threading.Tasks;
 using Cortex.Cryptography;
 using Microsoft.Extensions.Options;
-using Nethermind.BeaconNode.MockedStart;
-using Nethermind.BeaconNode.Services;
 using Nethermind.Core2;
+using Nethermind.Core2.Configuration;
 using Nethermind.Core2.Crypto;
 using Nethermind.Core2.Types;
 using Nethermind.HonestValidator.Services;
