+++ src/Nethermind/Nethermind.HonestValidator/MockedStart/QuickStartServiceCollectionExtensions.cs
@@ -19,14 +19,12 @@ using System.Linq;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
-using Nethermind.BeaconNode;
-using Nethermind.BeaconNode.MockedStart;
-using Nethermind.BeaconNode.Services;
 using Nethermind.Core2;
 using Nethermind.Core2.Types;
 using Nethermind.HonestValidator.Services;
 using Nethermind.Logging.Microsoft;
 using Nethermind.Core2.Configuration;
+using Nethermind.Core2.Configuration.MockedStart;
 
 namespace Nethermind.HonestValidator.MockedStart
 {
