+++ src/Nethermind/Nethermind.HonestValidator/Nethermind.HonestValidator.csproj
@@ -11,6 +11,7 @@
     </PropertyGroup>
 
     <ItemGroup>
+        <PackageReference Include="Cortex.Cryptography.Bls" Version="0.4.0" />
         <PackageReference Include="Microsoft.Extensions.Configuration.Binder" Version="3.0.1" />
         <PackageReference Include="Microsoft.Extensions.DependencyInjection.Abstractions" Version="3.0.1" />
         <PackageReference Include="Microsoft.Extensions.Hosting.Abstractions" Version="3.0.1" />
@@ -29,7 +30,6 @@
         <ProjectReference Include="..\Nethermind.Core2.Configuration\Nethermind.Core2.Configuration.csproj" />
         <ProjectReference Include="..\Nethermind.Core2.Json\Nethermind.Core2.Json.csproj" />
         <ProjectReference Include="..\Nethermind.Logging.Microsoft\Nethermind.Logging.Microsoft.csproj" />
-        <ProjectReference Include="..\Nethermind.BeaconNode\Nethermind.BeaconNode.csproj" />
     </ItemGroup>
 
 </Project>
